LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-traverse-navigation-no-bfcache.https-expected.txt

CONSOLE MESSAGE: SyntaxError: The string did not match the expected pattern.

Harness Error (TIMEOUT), message = null

TIMEOUT pageswap on traverse navigation from script Test timed out
